Homework 9.1.23: Fixed Deleting With Sorting
Fixed the deleteFolders function in folderStore to accommodate being given an array of folderIDs to delete. Finished deleteFolder to properly construct an array containing the correct folderIDS to be removed. It then calls folderStore.deleteFolder with those IDS. Added deleteFolder2 to do the same thing except when the sorting mode is not league.